% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{toSentence}
\alias{toSentence}
\title{Finding a sentence element among ancestors any element in a timeml tree}
\usage{
toSentence(descendent)
}
\arguments{
\item{descendent}{Any timeml element, presumed to be part of a timeml tree}
}
\value{
The sentence element that dominates the input timeml element, if a unique such sentence element exists; otherwise an error message.
}
\description{
\code{toSentence} takes an element from a timeml tree and determines whether there is a unique sentence element \code{<s>...</s>} among its ancestors.  If so, this sentence element is returned; if not, an error is printed.
}
\details{
TimeBank 1.2 does not include any sentence elements by default.  In order for this function to successfully detect sentences, the timeml tree hosting the input element must have been created from the documents in \code{<path>/timebank_1_2/data/extra}.  In the examples below, \code{docs} is a timeml tree.
}
\examples{
toSentence(sidToSignal("d119s96", docs)) 
toSentence(eidToEvent("d4e7", docs))
}
\seealso{
\code{printSentence}
}
